pyproject.toml
[project]
name = "redmine-mcp"
version = "0.1.0"
description = "Redmine MCP server for Claude Code - manage Redmine issues via MCP tools"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Jack Tsai", email = "snowild@gmail.com"},
]
maintainers = [
    {name = "Jack Tsai", email = "snowild@gmail.com"},
]
keywords = ["redmine", "mcp", "claude", "project-management", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Groupware",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "mcp[cli]>=1.9.4",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/your-username/redmine-mcp"
"Bug Reports" = "https://github.com/your-username/redmine-mcp/issues"
"Source" = "https://github.com/your-username/redmine-mcp"
"Documentation" = "https://github.com/your-username/redmine-mcp/blob/main/README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]

[project.scripts]
redmine-mcp = "redmine_mcp.server:main"
